Fix CurveNode.length inaccuracies (#764).

+++ src/player/CurveNode.h
@@ -68,7 +68,7 @@ class AVG_API CurveNode : public VectorNode
         float getTexCoord2() const;
         void setTexCoord2(float tc);
 
-        int getCurveLen() const;
+        float getCurveLen() const;
         glm::vec2 getPtOnCurve(float t) const;
 
         virtual void calcVertexes(const VertexDataPtr& pVertexData, Pixel32 color);
